LPS-104633 Moving to better places
