+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/servlet/taglib/ui/LayoutSEOScreenNavigationEntry.java
@@ -12,29 +12,27 @@
  * details.
  */
 
-package com.liferay.layout.admin.web.internal.servlet.taglib;
+package com.liferay.layout.admin.web.internal.servlet.taglib.ui;
 
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.layout.admin.web.internal.constants.LayoutScreenNavigationEntryConstants;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutConstants;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorCategory;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorCategoryUtil;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntryUtil;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.util.AggregateResourceBundle;
-import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.io.IOException;
 
-import java.util.List;
 import java.util.Locale;
+import java.util.Objects;
 import java.util.ResourceBundle;
-import java.util.stream.Stream;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -45,8 +43,11 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Alejandro Tardín
  */
-@Component(service = ScreenNavigationEntry.class)
-public class LayoutGeneralScreenNavigationEntry
+@Component(
+	property = "screen.navigation.entry.order:Integer=1",
+	service = ScreenNavigationEntry.class
+)
+public class LayoutSEOScreenNavigationEntry
 	implements ScreenNavigationEntry<Layout> {
 
 	@Override
@@ -56,7 +57,7 @@ public class LayoutGeneralScreenNavigationEntry
 
 	@Override
 	public String getEntryKey() {
-		return LayoutScreenNavigationEntryConstants.ENTRY_KEY_GENERAL;
+		return LayoutScreenNavigationEntryConstants.ENTRY_KEY_SEO;
 	}
 
 	@Override
@@ -72,18 +73,24 @@ public class LayoutGeneralScreenNavigationEntry
 
 	@Override
 	public boolean isVisible(User user, Layout layout) {
-		List<FormNavigatorCategory> formNavigatorCategories =
-			FormNavigatorCategoryUtil.getFormNavigatorCategories(
-				FormNavigatorConstants.FORM_NAVIGATOR_ID_LAYOUT);
+		Group group = _groupLocalService.fetchGroup(layout.getGroupId());
 
-		Stream<FormNavigatorCategory> formNavigatorCategoryStream =
-			formNavigatorCategories.stream();
+		if ((group != null) && group.isLayoutPrototype()) {
+			return false;
+		}
 
-		return formNavigatorCategoryStream.anyMatch(
-			category -> ListUtil.isNotEmpty(
-				FormNavigatorEntryUtil.getFormNavigatorEntries(
-					FormNavigatorConstants.FORM_NAVIGATOR_ID_LAYOUT,
-					category.getKey(), user, layout)));
+		Layout draftLayout = _layoutLocalService.fetchLayout(
+			_portal.getClassNameId(Layout.class), layout.getPlid());
+
+		if ((Objects.equals(layout.getType(), LayoutConstants.TYPE_CONTENT) ||
+			 Objects.equals(
+				 layout.getType(), LayoutConstants.TYPE_ASSET_DISPLAY)) &&
+			(draftLayout == null)) {
+
+			return false;
+		}
+
+		return true;
 	}
 
 	@Override
@@ -94,7 +101,7 @@ public class LayoutGeneralScreenNavigationEntry
 
 		_jspRenderer.renderJSP(
 			httpServletRequest, httpServletResponse,
-			"/edit_layout_general.jsp");
+			"/layout/screen/navigation/entries/seo.jsp");
 	}
 
 	protected ResourceBundle getResourceBundle(Locale locale) {
@@ -106,7 +113,13 @@ public class LayoutGeneralScreenNavigationEntry
 	}
 
 	@Reference
+	private GroupLocalService _groupLocalService;
+
+	@Reference
 	private JSPRenderer _jspRenderer;
+
+	@Reference
+	private LayoutLocalService _layoutLocalService;
 
 	@Reference
 	private Portal _portal;
